requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.4.2
pandas==3.0.0
psycopg2-binary==2.9.11
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
rich==14.3.2
shellingham==1.5.4
six==1.17.0
typer==0.21.1
typing_extensions==4.15.0
urllib3==2.6.3
